+++ tilebox-datasets/tilebox/datasets/protobuf_conversion/protobuf_xarray.py
@@ -231,7 +231,7 @@ class _SimpleFieldConverter(_FieldConverter):
         elif buffer_size > len(self._data):
             # resize the data buffer to the new capacity, by just padding it with zeros at the end
             missing = buffer_size - len(self._data)
-            self._data = np.pad(  # ty: ignore[no-matching-overload]
+            self._data = np.pad(
                 self._data,
                 ((0, missing), (0, 0)),
                 constant_values=self._type.fill_value,
@@ -309,7 +309,7 @@ class _ArrayFieldConverter(_FieldConverter):
         else:  # resize the data buffer to the new capacity, by just padding it with zeros at the end
             missing_capacity = self._capacity - self._data.shape[0]
             missing_array_dim = self._array_dim - self._data.shape[1]
-            self._data = np.pad(  # ty: ignore[no-matching-overload]
+            self._data = np.pad(
                 self._data,
                 ((0, missing_capacity), (0, missing_array_dim), (0, 0)),
                 constant_values=self._type.fill_value,